+++ lib/getch/command.rb
@@ -4,14 +4,22 @@ module Getch
   class Command
     def initialize(cmd)
       @cmd = cmd
-      @block_size = 512
+      @block_size = 1024
+      @log = Getch::Log.new
     end
 
     def run!
-      puts "Running command: " + @cmd.gsub(/\"/, '')
+      @log.info "Running command: " + @cmd.gsub(/\"/, '')
 
-      Open3.popen3(@cmd) do |stdin, stdout, stderr|
+      Open3.popen3(@cmd) do |stdin, stdout, stderr, wait_thr|
         stdin.close_write
+        code = wait_thr.value
+
+        # only stderr
+        begin
+          @logger.error stderr.readline until stderr.eof.nil?
+        rescue EOFError
+        end
 
         begin
           files = [stdout, stderr]
@@ -24,25 +32,14 @@ module Getch
               # writable = ready[1]
               # exceptions = ready[2]
 
-              readable.each do |f|
-                fileno = f.fileno
-
-                begin
-                  data = f.read_nonblock(@block_size)
-
-                  # Do something with the data...
-                  puts "#{data}" if DEFAULT_OPTIONS[:verbose]
-                rescue EOFError
-                  puts "fileno: #{fileno} EOF"
-                end
-              end
+              display_lines(readable)
             end
           end
         rescue IOError => e
           puts "IOError: #{e}"
         end
+        @logger.debug "Done - #{@cmd} - #{code}"
       end
-      puts "Done"
     end
 
     private
@@ -50,6 +47,19 @@ module Getch
     # Returns true if all files are EOF
     def all_eof(files)
       files.find { |f| !f.eof }.nil?
+    end
+
+    def display_lines(block)
+      block.each do |f|
+        begin
+          data = f.read_nonblock(@block_size)
+          puts "#{data}" if DEFAULT_OPTIONS[:verbose]
+        rescue EOFError
+          puts ""
+        rescue => e
+          puts "Fatal - #{e}"
+        end
+      end
     end
   end
 end